M0-212 feat: created the data collector for exercice and sleep options and saving in the core data

+++ Mini06-watchOSApp WatchKit App/CoreData/Exercice/Exercice+CoreDataProperties.swift
@@ -18,7 +18,7 @@ extension Exercice {
 
     @NSManaged public var countSteps: Int64
     @NSManaged public var didExercice: Bool
-    @NSManaged public var distanceWalked: Float
+    @NSManaged public var distanceWalked: Double
     @NSManaged public var id: UUID?
     @NSManaged public var kcalLost: Int64
     @NSManaged public var bpm: NSSet?
